Add listener, set state
